HOVER implemented, improved names, organized code, fixed/improved some methods
Hover:
Lower elevator should never drift down from its weight anymore. Hover() is a boolean that is either active or not, and it has to be turned on and off by changing the hover boolean. I may have missed some spots where hover should be turned on or off. If hover is turned on and a button doesnt turn it off before acting, that button will not be able to control lower elevator, so it is important to watch out for that.

Naming system:
instances in which "position" was used now use "destination" when appropriate. This includes
- setPosition -> sendToDestination
- targetPosition -> targetDestination
- addToPosition -> addToDestination
- setPositionRunning -> sendCombinedRunning

Prints:
Commented out prints that could potentially clog log, and added some that could be useful for testing

Comments:
Added comments to describe each method's function

Variables: Minimum values changed to 0 because they *should* be 0. However, it seems like the encoders get thrown off when the elevator reaches the top, and it is likely that the bot will slam into the bottom when going back down. Tinker with min values to add a cushion until we figure out a solution.

Methods:
sendCombinedToDestination: Has been drastically altered. Here are some problems with the old version
- near case only ran if the destination (upper and lower combined) was going to somewhere below lower near min. It makes more sense to make it run if lower's destination alone is nearing lower min.
- It would run the near case if elevator was moving from 0 to 100,000 (upwards), which is when it has no danger of slamming bottom
- if lower tripped lim switch unexpectedly, both lower AND upper would stop
- code was confusing, and had if with a condition that would never occur when inside the while loop because the while only ran when that condition wasnt true
- would print things that weren't true in some cases, like the "cannot reach target position"

New sendCombined:
- Checks for upper and lower near cases SEPERATELY
- Only runs lower near if lower is GOING DOWN and NEARING BOTTOM
- Also has an upper near just to be safe
- Seperated upper and lower so they do not influence each other when they shouldn't
- fixed print statement

addToDestination: code hadnt been completed, but is now written. However, it might add to destination every single refresh that button is pressed, which would make robot want to go to very top. It will not add if it is sending it above the max though.

Commented code:
I realized that we had duplicate code, so I cleared it. Check the removed commented code if you think it should be added back, but they were methods like "upper elevator basic movement w/o limits" which serves the exact same function as manual fored upper elevator movement.

This code has been thoroughly checked through multiple times (I spent like 6 hours on this) but it probably has some basic bugs like mixing "upper" with "lower", and that doesn't include the fact that it might just not work. It definitely needs testing.